requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
PyYAML==6.0.2
requests==2.32.3

